requirements.txt
streamlit==1.32.0
python-dotenv==1.0.1
google-search-results==2.4.2
requests==2.31.0
openai==1.12.0
pandas==2.2.1
